+++ src/main/java/ladder/Model/Ladder.java
@@ -5,11 +5,10 @@ import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 
 public class Ladder {
-    private static final int MIN_WIDTH = 0;
-
     private final List<LadderLine> ladderLines = new ArrayList<>();
 
     private final MakeLineStrategy makeLineStrategy = new MakeLineStrategyRandom();
+    private final Move move = new Move();
     private final ResultMap resultMap = new ResultMap();
 
     public Ladder(List<String> attendances, int height) {
@@ -27,34 +26,11 @@ public class Ladder {
 
         ladderLines.forEach(
                 ladderLine -> {
-                    positionX.set(moveDownJustOnce(ladderLine, positionX.get()));
+                    positionX.set(move.down(ladderLine, positionX.get()));
                 }
         );
 
         resultMap.put(name, positionX.get());
-    }
-
-    private int moveDownJustOnce(LadderLine ladderLine, int index) {
-        int result = index;
-        final int width = ladderLine.size();
-
-        if (index - 1 >= MIN_WIDTH && isExistHorizonLines(ladderLine.horizonLines(), index - 1)) {
-            return result - 1;
-        }
-
-        if (index + 1 < width && isExistHorizonLines(ladderLine.horizonLines(), index)) {
-            return result + 1;
-        }
-
-        return result;
-    }
-
-    private boolean isExistHorizonLines(List<Boolean> line, int index) {
-        if (index >= line.size()) {
-            return false;
-        }
-
-        return line.get(index);
     }
 
     public ResultMap resultMap() {
