+++ src/main/java/ladder/View/InputView.java
@@ -92,6 +92,7 @@ public class InputView {
         System.out.println("");
         System.out.println("실행 결과");
         System.out.println(results.get(key));
+        System.out.println("");
     }
 
 }
